Created Databases and software developments

+++ Core/MultiStoreIntegration.Domain/Entities/Stock.cs
@@ -1,9 +1,4 @@
 ﻿using MultiStoreIntegration.Domain.Entities.common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MultiStoreIntegration.Domain.Entities
 {
@@ -15,7 +10,7 @@ namespace MultiStoreIntegration.Domain.Entities
         public string Size { get; set; }
         public string Color { get; set; }
         public int Quantity { get; set; }
-        public int UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
 
         // Navigation properties
         public ICollection<Sale> Sales { get; set; }
